requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
abb-librws=1.2.0=pypi_0
absl-py=1.4.0=pypi_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
autolab-core=1.1.1=pypi_0
autolab-perception=1.1.0=dev_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
binutils_impl_linux-64=2.40=hf600244_0
black=21.4b2=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.1=pypi_0
cairo=1.16.0=hcf35c78_1003
certifi=2023.7.22=py38h06a4308_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=2.2.1=pyhd8ed1ab_0
cmake=3.25.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
colorlog=6.7.0=py38h578d9bd_1
comm=0.1.4=pyhd8ed1ab_0
cryptography=41.0.2=py38h774aba0_0
cycler=0.11.0=pyhd8ed1ab_0
cython=3.0.0=pypi_0
cytoolz=0.12.2=py38h01eb140_0
dask-core=2021.11.2=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=he372182_0
debugpy=1.6.8=py38h17151c0_0
decorator=5.1.1=pyhd8ed1ab_0
detectron2=0.6=dev_0
dill=0.3.7=pyhd8ed1ab_0
efficientnet-pytorch=0.7.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
enum34=1.1.10=py38h32f6830_2
executing=1.2.0=pyhd8ed1ab_0
expat=2.4.8=h27087fc_0
ffmpeg=4.3=hf484d3e_0
ffmpeg-python=0.2.0=pypi_0
filelock=3.9.0=py38h06a4308_0
fontconfig=2.14.0=h8e229c2_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.6.0=pyh1a96a4e_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20221221=pyhd8ed1ab_0
gcc=12.1.0=h9ea6d83_10
gcc_impl_linux-64=12.1.0=hea43390_17
geos=3.10.2=h9c3ff4c_0
gettext=0.19.8.1=hf34092f_1004
giflib=5.2.1=h5eee18b_3
glib=2.66.1=h680cd38_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.57.0=pypi_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_101
huggingface-hub=0.16.4=pypi_0
hydra-core=1.3.2=pypi_0
icu=64.2=he1b5a44_1
idna=3.4=py38h06a4308_0
imagecodecs-lite=2019.12.3=py38h26c90d9_6
imageio=2.31.1=pyh24c5eb1_0
imgaug=0.4.0=pyhd8ed1ab_1
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.8=pypi_0
ipykernel=6.25.1=pyh71e2992_0
ipython=8.12.2=pypi_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=4.12.0=py38h578d9bd_0
kernel-headers_linux-64=2.6.32=he073ed8_16
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=9.0.1=default_hb4e5071_5
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-devel_linux-64=12.1.0=h1ec3361_17
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=13.1.0=he5830b7_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm9=9.0.1=default_hc23dcda_4
libopencv=4.4.0=py38_2
libosqp=0.6.3=h59595ed_0
libpng=1.6.39=h5eee18b_0
libqdldl=0.1.5=h27087fc_1
libsanitizer=12.1.0=ha89aaad_17
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=13.1.0=hfd8a6a1_0
libtasn1=4.19.0=h5eee18b_0
libtiff=4.2.0=hf544144_3
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.13=h7f98852_1004
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lit=15.0.7=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h2af1d28_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multiprocess=0.70.15=py38h01eb140_0
munch=4.0.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.32=h9c3ff4c_1
nss=3.59=h2c00c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pyhd8ed1ab_0
opencv=4.4.0=py38_2
opencv-python=4.8.0.76=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1v=h7f8727e_0
osqp=0.6.3=py38h01efb38_1
packaging=23.1=pypi_0
pandas=1.4.4=py38h6a678d5_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.4.0=pyhd8ed1ab_0
pathlib=1.0.1=py38h578d9bd_7
pathspec=0.11.2=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh1a96a4e_2
phoxipy=0.0.1=dev_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_0
pip=23.2.1=py38h06a4308_0
pixman=0.38.0=h516909a_1003
portalocker=1.4.0=py_0
pretrainedmodels=0.7.4=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.24.2=pypi_0
psutil=5.9.5=py38h1de0b5d_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-opencv=4.4.0=py38h23f93f0_2
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybind11=2.11.1=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyserial=3.5=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fcl=0.7.0.5=pypi_0
python_abi=3.8=2_cp38
pytorch-mutex=1.0=cpu
pytz=2022.7=py38h06a4308_0
pywavelets=1.4.1=py38h7e4f40d_0
pyyaml=5.1.2=py38h516909a_1
pyzmq=25.1.1=py38h509eb50_0
qdldl-python=0.1.5.post2=py38h8f669ce_0
qt=5.12.5=hd8c4c69_1
readline=7.0=h7b6447c_5
regex=2023.8.8=pypi_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
ruamel.yaml=0.17.32=py38h01eb140_0
ruamel.yaml.clib=0.2.7=py38h1de0b5d_1
safetensors=0.3.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.2.0=py38h6a678d5_1
scipy=1.7.3=pypi_0
segmentation-models-pytorch=0.3.3=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=59.8.0=py38h578d9bd_1
shapely=1.8.1=py38h97f7145_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pyh04b8f61_3
sysroot_linux-64=2.12=he073ed8_16
tabulate=0.9.0=pyhd8ed1ab_1
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.3.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.7.10=pypi_0
timm=0.9.2=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pyhd8ed1ab_0
torch=2.0.1+cu118=pypi_0
torchaudio=2.0.2+cu118=pypi_0
torchvision=0.15.2+cu118=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.66.1=pyhd8ed1ab_0
tracikpy=0.2.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
trimesh=3.23.1=pyhd8ed1ab_0
triton=2.0.0=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
urllib3=1.26.16=py38h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py38h06a4308_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
yumiplanning=0.1.0=dev_0
yumirws=0.0.1=dev_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
